+++ package.json
@@ -16,6 +16,7 @@
     "child_process": "^1.0.2",
     "discord.js": "^11.1.0",
     "discord.js-commando": "^0.9.0",
+    "mongodb": "^2.2.31",
     "request": "^2.81.0",
     "zerorpc": "^0.9.7"
   },
